Last commit git bugged so here is the fix

+++ dock/startservs.sh
@@ -2,8 +2,10 @@
 # starts the services required to run stuff (services in the container can't be autostarted)
 
 set -eu
-mv -n /defmysql/* /var/lib/mysql/
-mkdir -p /defmysql
+if [ -f "/defmysql/" ];
+then
+	mv -n /defmysql/* /var/lib/mysql/
+fi;
 service mariadb start
 service apache2 start
 sleep 2147483647
